Added readme and fixed install script.

+++ skiros2/skiros2/scripts/install_fd_task_planner.sh
@@ -6,28 +6,28 @@ echo "Installing planner..."
 
 #Navigate to install folder
 folder="$HOME/utils"
-# echo "Select install folder or leave blank for default [default: $default]:"
-# read folder
-# if [[ $folder == "" ]]; then
-#     folder=${default/"~"/$HOME}
-# else
-#     folder=${folder/"~"/$HOME}
-# fi
 mkdir -p $folder
 cd $folder
-# echo $folder
 
-#Install
-wget "http://gki.informatik.uni-freiburg.de/tools/tfd/downloads/version-0.4/${tfd}.tgz"
-tar xzf "${tfd}.tgz"
-cd "${tfd}"
-sed -e s/"-Werror"//g -i ./downward/search/Makefile
-sed -e s/"translate\/"//g -i ./downward/plan.py
-sed -e s/"preprocess\/"//g -i ./downward/plan.py
-sed -e s/"search\/"//g -i ./downward/plan.py
-./build
-cd - >/dev/null
-rm -r "${tfd}.tgz"
+# Installing planner
+if [ ! -d "$folder/${tfd}" ]; then
+    echo "Installing FD planner in $folder"
+    mkdir -p $folder
+    cd $folder
+    echo $folder
+    wget "http://gki.informatik.uni-freiburg.de/tools/tfd/downloads/version-0.4/${tfd}.tgz"
+    tar xzf "${tfd}.tgz"
+    cd "${tfd}"
+    sed -e s/"-Werror"//g -i ./downward/search/Makefile
+    sed -e s/"translate\/"//g -i ./downward/plan.py
+    sed -e s/"preprocess\/"//g -i ./downward/plan.py
+    sed -e s/"search\/"//g -i ./downward/plan.py
+    ./build
+    cd -
+    rm -r "${tfd}.tgz"
+else
+    echo "Folder $folder/${tfd} already exists. Skipping installation."
+fi
 
 #Add environment variable to bashrc
 export TFD_HOME=$(pwd)
@@ -39,7 +39,7 @@ if [ -z "$temp" ]; then
 	echo $string2 >> ~/.bashrc
 	echo "Printing string in .bashrc: $string"
 else
-	echo "No export string added to bashrc (it is already there)"
+	echo "Found planner path in bashrc: $string. Skip printing."
 fi
 
 # source ~/.bashrc
